redhat-rpm-config: change default arch level target for all ARM v7 to armv7hl
All ARM version 7 systems support a vector hardware floating point unit and
have the ability to run using the hard floating point ABI (aapcs-vfpv3-d16).
This is the only configuration we support as a v7 target, so we force the
use of hard floating point. This prevents e.g. packages being built with
a armv5tel target on an armv7 system without explicit intent.

Signed-off-by: Jon Masters <jcm@jonmasters.org>